+++ example/flutter_counter/lib/main.dart
@@ -32,14 +32,14 @@ class _Counter with ReactiveComponent {
   /// will be conciser as below,
   ///
   /// '''dart
-  /// late final increment = VoidReactiveSink((_) {
+  /// late final increment = VoidReactiveSink(() {
   ///        _count.data++;
   ///      }, disposer: disposer);
   /// '''
   ///
   /// For more information, see [ReactiveSink]'s API documentation.
   VoidReactiveSink _increment;
-  VoidReactiveSink get increment => _increment ??= VoidReactiveSink((_) {
+  VoidReactiveSink get increment => _increment ??= VoidReactiveSink(() {
         // Increments _count on a increment event is delivered.
         _count.data++;
       }, disposer: disposer);